+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Change log
 haze uses [PVP Versioning][1].
 The change log is available [on GitHub][2].
 
+0.1.1
+====
+* [51](https://github.com/cronokirby/haze/issues/51)
+
+    Fix a bug where the announcer would log messages in a constant loop
+    after running out of trackers.
+
 0.1.0
 =====
 * Initially created.